Move api related functions to api.js - Update api function imports - Update api functions to be async
